backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
alembic==1.12.1
celery==5.3.6
redis==5.0.1
bcrypt==4.0.1
python-dotenv==1.0.0
pymysql==1.1.0
psycopg2-binary==2.9.9
requests==2.31.0
